+++ ci/script.sh
@@ -1,15 +1,25 @@
 set -euxo pipefail
 
 main() {
+    cargo check --target $TARGET
+
+    if [ $TRAVIS_RUST_VERSION = nightly ]; then
+        cargo check --target $TARGET --features inline-asm
+    fi
+
     case $TARGET in
         thumbv7em-none-eabi*)
             cargo check --target $TARGET --features cm7-r0p1
-            cargo check --target $TARGET
+
+            if [ $TRAVIS_RUST_VERSION = nightly ]; then
+                cargo check --target $TARGET --features 'cm7-r0p1 inline-asm'
+            fi
             ;;
+
         thumbv*-none-eabi*)
-            cargo check --target $TARGET
             ;;
-        *)
+
+        x86_64-unknown-linux-gnu)
             cargo test --target $TARGET
             ;;
     esac